back/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cachetools>=6.0.0",
    "fastapi[standard]>=0.115.12",
    "python-dotenv>=1.1.0",
    "tortoise-orm>=0.25.0",
]
